# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API CMS",
        "description": "Todas las rutas de la API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/v1/pistas": {
            "get": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Obtener una lista de pistas o horarios disponibles",
                "operationId": "221208cda654433fb35877dede189336",
                "parameters": [
                    {
                        "name": "fecha",
                        "in": "query",
                        "description": "Fecha para consultar horarios disponibles",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-09-21 10:00:00"
                        }
                    },
                    {
                        "name": "deporte",
                        "in": "query",
                        "description": "Nombre del deporte para filtrar pistas",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Fútbol"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de pistas o horarios obtenidos con éxito"
                    },
                    "404": {
                        "description": "No se han encontrado las pistas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Crear una nueva pista",
                "operationId": "4345632f15c1a55bb270a7a2215ae5d5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "deporte_id",
                                    "longitud",
                                    "ancho"
                                ],
                                "properties": {
                                    "deporte_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "longitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 30.5
                                    },
                                    "ancho": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pista creada con éxito"
                    },
                    "403": {
                        "description": "No tiene permisos para realizar esta acción"
                    },
                    "404": {
                        "description": "Deporte no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/pistas/{id}": {
            "get": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Obtener detalles de una pista específica",
                "operationId": "474be561d7de8dff70a01214dfe596e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la pista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la pista obtenidos con éxito"
                    },
                    "404": {
                        "description": "Pista no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Actualizar los datos de una pista",
                "operationId": "fb3ccef734f1004910babede4b6bdb7f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la pista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "deporte_id",
                                    "longitud",
                                    "ancho"
                                ],
                                "properties": {
                                    "deporte_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "longitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 30.5
                                    },
                                    "ancho": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pista actualizada con éxito"
                    },
                    "404": {
                        "description": "Pista no encontrada"
                    },
                    "403": {
                        "description": "No tiene permisos para realizar esta acción"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pistas"
                ],
                "summary": "Eliminar una pista",
                "operationId": "abfe46ac0d562e2b3803c19d3c6f71e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la pista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pista eliminada con éxito"
                    },
                    "404": {
                        "description": "Pista no encontrada"
                    },
                    "403": {
                        "description": "No tiene permisos para realizar esta acción"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usa un token de acceso para autenticar las peticiones",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}